+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Minor fix / optimization in the `RegExp` constructor (NCG and `dotAll`) polyfill
 - Added some more workarounds for a Safari < 13 bug with silent ignore of non-writable array `.length`
 - Added detection of a Webkit [bug](https://bugs.webkit.org/show_bug.cgi?id=297532): `Iterator.prototype.flatMap` throws on iterator without `return` method
+- Added detection of a V8 ~ Chromium < 144 [bug](https://issues.chromium.org/issues/454630441): `Uint8Array.prototype.setFromHex` throws an error on length-tracking views over ResizableArrayBuffer
 - Compat data improvements:
   - [`Map` upsert proposal](https://github.com/tc39/proposal-upsert) features marked as [shipped in V8 ~ Chrome 145](https://issues.chromium.org/issues/434977728#comment4)
   - [Joint iteration proposal](https://github.com/tc39/proposal-joint-iteration) features marked as [shipped in FF148](https://bugzilla.mozilla.org/show_bug.cgi?id=2003333#c8)
@@ -21,6 +22,7 @@
   - Added Opera Android [93](https://forums.opera.com/topic/87267/opera-for-android-93) and [94](https://forums.opera.com/topic/87678/opera-for-android-94) compat data mapping
   - Added Electron 41 compat data mapping
   - `Iterator.prototype.flatMap` marked as supported from Safari 26.2 and Bun 1.2.21 because of a [bug](https://bugs.webkit.org/show_bug.cgi?id=297532): throws on iterator without `return` method
+  - `Uint8Array.prototype.setFromHex` marked as supported from V8 ~ Chromium 144 because of a [bug](https://issues.chromium.org/issues/454630441): throws an error on length-tracking views over ResizableArrayBuffer
 
 ### [3.47.0 - 2025.11.18](https://github.com/zloirock/core-js/releases/tag/v3.47.0)
 - Changes [v3.46.0...v3.47.0](https://github.com/zloirock/core-js/compare/v3.46.0...v3.47.0) (117 commits)